+++ tests/rss/test_rss_utils.py
@@ -1,11 +1,18 @@
 import os 
 os.environ["OMP_NUM_THREADS"] = "1"
 
-from autoplex.data.rss.utils import HookeanRepulsion
 from quippy.potential import Potential
-from autoplex.fitting.common.utils import extract_gap_label
 from ase.io import read
 import numpy as np
+import pytest
+from autoplex.data.common.utils import (cur_select, 
+                                        boltzhist_cur_one_shot, 
+                                        ElementCollection, 
+                                        boltzhist_cur_dual_iter)
+from autoplex.data.rss.utils import CustomPotential, extract_pairstyle
+from autoplex.fitting.common.utils import extract_gap_label
+from autoplex.data.rss.utils import HookeanRepulsion, process_rss, minimize_structures
+from pymatgen.io.ase import AseAtomsAdaptor
 
 
 def test_hookean(test_dir, memory_jobstore):
